requirements.txt
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
mysqlclient==2.1.1
numpy==1.23.3
orjson==3.8.0
pandas==1.4.4
passlib==1.7.4
pycares==4.2.2
pycparser==2.21
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
scipy==1.9.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.40
uvicorn==0.17.6
